requirements.txt
streamlit
dropbox
openai
python-docx
pymupdf
